Move common util functions to same folder
